fork_sync_data: Display commits which can be cherry-picked
This commit adds a column "Supports clean cherry-pick" to the
upstream-only view. This will make it more clear if it
is simple or hard to cherry-pick a bugfix commit.

Note: Because many commits depend on each-other it is often the case
that a commit cannot be cleanly cherry-picked to the downstream
repo.